+++ src/Middleware/IdFilter.js
@@ -9,17 +9,20 @@ class IdFilter {
     this.routeHelper = routeHelper
   }
 
-  async handle ({ request, params }, next) {
-    this._setMatchedUrl(request, params)
-    this._setParentColumns(request)
+  async handle (ctx, next) {
+    this._setMatchedUrl(ctx.request, ctx.params)
+    this._setParentColumns(ctx.request)
+
+    // We should call custom middlewares in here if there is any
+    await this._callCustomMiddlewares(ctx)
 
     // Fetching idKey data automatically
-    for (const idKey of this._getSpecialIdKeys(request.apix.url)) {
+    for (const idKey of this._getSpecialIdKeys(ctx.request.apix.url)) {
       const Middleware = this.routeHelper.getMiddlewareModel(idKey)
       const Model = this._loadModel(Middleware.model)
 
       try {
-        request.apix.layers[pluralize.singular(Middleware.table)] = await this._findOrFail(Model, params, idKey)
+        ctx.request.apix.layers[pluralize.singular(Middleware.table)] = await this._findOrFail(Model, ctx.params, idKey)
       } catch (error) {
         console.log(error)
         throw new HttpException(404, `Record not found on ${capitalCase(Middleware.model)}.`)
@@ -27,6 +30,54 @@ class IdFilter {
     }
 
     await next()
+  }
+
+  async _callCustomMiddlewares (ctx) {
+    return new Promise((resolve) => {
+      const Model = this.routeHelper.get(ctx.request.apix.url)
+
+      if (Model.middlewares.length === 0) {
+        return resolve()
+      }
+
+      this.currentMiddlewareIndex = 0
+      this._callNextMidleware(ctx, Model, () => {
+        resolve()
+      })
+    })
+  }
+
+  _callNextMidleware (ctx, Model, next) {
+    if (this.currentMiddlewareIndex >= Model.middlewares.length) {
+      return next()
+    }
+
+    const item = Model.middlewares[this.currentMiddlewareIndex++]
+
+    if (this._hasMiddleware(ctx, item)) {
+      const MiddlewareClass = this._getMiddlewareInstance(item)
+      const instance = new MiddlewareClass()
+      instance.handle(ctx, () => {
+        this._callNextMidleware(ctx, Model, next)
+      })
+    } else {
+      next()
+    }
+  }
+
+  _getMiddlewareInstance (item) {
+    if (typeof item === 'string') {
+      return use(item)
+    }
+    return use(item.middleware)
+  }
+
+  _hasMiddleware (ctx, item) {
+    if (typeof item === 'string') {
+      return true
+    }
+
+    return item.method === ctx.request.method()
   }
 
   _setMatchedUrl (request, params) {