+++ src/Models/XModel.js
@@ -4,6 +4,10 @@ class XModel extends Model {
   static get actions () {
     return ['GET', 'POST', 'PUT', 'DELETE']
   }
+
+  static get middlewares () {
+    return []
+  }
 }
 
 module.exports = XModel