+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Custom middleware feature has been added.
+- Bug fixed on route generator.
+
 ## 0.1.0
 
 - First version has been relased.